Product price should display decimal

+++ app/models/product.rb
@@ -30,14 +30,14 @@ class Product < ApplicationRecord
   end
 
   def format_price
-    self.sim_card_price_currency == 'GBP' ? self.sim_card_price.currency.symbol + '' + self.price.to_i.to_s : self.price.to_i.to_s + '' + self.sim_card_price.currency.symbol
+    self.sim_card_price_currency == 'GBP' ? self.sim_card_price.currency.symbol + '' + self.price.to_f.to_s : self.price.to_f.to_s + '' + self.sim_card_price.currency.symbol
   end
 
   def format_sim_card_price
-    self.sim_card_price_currency == 'GBP' ? self.sim_card_price.currency.symbol + '' + self.sim_card_price.to_i.to_s : self.sim_card_price.to_i.to_s + '' + self.sim_card_price.currency.symbol
+    self.sim_card_price_currency == 'GBP' ? self.sim_card_price.currency.symbol + '' + self.sim_card_price.to_f.to_s : self.sim_card_price.to_f.to_s + '' + self.sim_card_price.currency.symbol
   end
 
   def format_setup_price
-    self.sim_card_price_currency == 'GBP' ? self.sim_card_price.currency.symbol + '' + self.setup_price.to_i.to_s : self.setup_price.to_i.to_s + '' + self.sim_card_price.currency.symbol
+    self.sim_card_price_currency == 'GBP' ? self.sim_card_price.currency.symbol + '' + self.setup_price.to_f.to_s : self.setup_price.to_f.to_s + '' + self.sim_card_price.currency.symbol
   end
 end
